% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{get_excel_url}
\alias{get_excel_url}
\title{Format Excel data from a URL for input into R}
\usage{
get_excel_url(file)
}
\arguments{
\item{file}{The Excel (.xls or .xlsx) file to be read in from a URL.}
}
\description{
This is a helper function to assist in reading in Excel data from a URL.
There is an active PR for hadley/readxl (#77) to deal with this issue, 
but it hasn't been merged yet. So, this hack was written to be able to 
correctly read in Excel data from the web.
}
\examples{
get_excel_url()
}
\keyword{Excel}

