% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{reg_output}
\alias{reg_output}
\title{Format univariate regression output neatly}
\usage{
reg_output(dep, ind, data)
}
\arguments{
\item{dep}{Specifies the dependent variable name. Should be a string.}

\item{ind}{Specifies the independent variable name. Should be a string.}

\item{data}{Specifies the data frame from which the dep and ind variables come.}
}
\description{
This function formats the key pieces of a univariate regression for a 
lay audience. It does not, and is not intended, to replace robust 
regression output that would be valuable for a more advanced 
audience/analyst.
}
\examples{
reg_output()
}
\keyword{regression}

